Move SynchronizedRegistry class back, so qfapi doesn't crash
