+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/api/sync/RegistrySynchronization.java
@@ -26,7 +26,7 @@ import net.minecraft.registry.SimpleRegistry;
 import net.minecraft.util.Identifier;
 
 import org.quiltmc.qsl.registry.impl.sync.registry.RegistryFlag;
-import org.quiltmc.qsl.registry.impl.sync.registry.SynchronizedRegistry;
+import org.quiltmc.qsl.registry.impl.sync.SynchronizedRegistry;
 import org.quiltmc.qsl.registry.impl.sync.server.ExtendedConnectionClient;
 
 /**
