+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/impl/sync/RegistrySyncInitializer.java
@@ -23,7 +23,6 @@ import net.minecraft.registry.Registries;
 import org.quiltmc.loader.api.ModContainer;
 import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
 import org.quiltmc.qsl.registry.impl.sync.modprotocol.ModProtocolImpl;
-import org.quiltmc.qsl.registry.impl.sync.registry.SynchronizedRegistry;
 import org.quiltmc.qsl.registry.impl.sync.server.ServerRegistrySync;
 
 @ApiStatus.Internal
