+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/impl/sync/client/ClientFabricRegistrySync.java
@@ -42,7 +42,7 @@ import org.quiltmc.qsl.networking.api.PacketByteBufs;
 import org.quiltmc.qsl.networking.api.PacketSender;
 import org.quiltmc.qsl.networking.api.client.ClientPlayNetworking;
 import org.quiltmc.qsl.registry.impl.sync.server.ServerFabricRegistrySync;
-import org.quiltmc.qsl.registry.impl.sync.registry.SynchronizedRegistry;
+import org.quiltmc.qsl.registry.impl.sync.SynchronizedRegistry;
 
 /**
  * Legacy (Fabric) registry sync.
