+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/impl/sync/registry/RegistryFlag.java
@@ -20,6 +20,7 @@ import org.jetbrains.annotations.ApiStatus;
 
 import net.minecraft.registry.SimpleRegistry;
 import net.minecraft.util.Identifier;
+import org.quiltmc.qsl.registry.impl.sync.SynchronizedRegistry;
 
 /**
  * Registry flags used to determine how registry (entries) should be handled by sync.
